# Add failure detection with higher granularity for read replicas