--- a/service/storage_proxy.cc
+++ b/service/storage_proxy.cc
@@ -3861,13 +3861,14 @@ db::read_repair_decision storage_proxy::new_read_repair_decision(const schema& s
         tracing::trace_state_ptr trace_state,
         const inet_address_vector_replica_set& preferred_endpoints,
         bool& is_read_non_local,
-        service_permit permit) {
+        service_permit permit,
+        clock_type::time_point timeout) {
     const dht::token& token = pr.start()->value().token();
     keyspace& ks = _db.local().find_keyspace(schema->ks_name());
     speculative_retry::type retry_type = schema->speculative_retry().get_type();
     gms::inet_address extra_replica;
 
-    inet_address_vector_replica_set all_replicas = get_live_sorted_endpoints(ks, token);
+    inet_address_vector_replica_set all_replicas = get_live_sorted_endpoints_for_read(ks, token, cl, timeout);
     // Check for a non-local read before heat-weighted load balancing
     // reordering of endpoints happens. The local endpoint, if
     // present, is always first in the list, as get_live_sorted_endpoints()
@@ -3997,6 +3998,7 @@ storage_proxy::query_singular(lw_shared_ptr<query::read_command> cmd,
     bool is_read_non_local = false;
 
     const auto tmptr = get_token_metadata_ptr();
+    auto timeout = query_options.timeout(*this);
     for (auto&& pr: partition_ranges) {
         if (!pr.is_singular()) {
             throw std::runtime_error("mixed singular and non singular range are not supported");
@@ -4009,7 +4011,7 @@ storage_proxy::query_singular(lw_shared_ptr<query::read_command> cmd,
 
         auto read_executor = get_read_executor(cmd, schema, std::move(pr), cl, repair_decision,
                                                query_options.trace_state, replicas, is_read_non_local,
-                                               query_options.permit);
+                                               query_options.permit, timeout);
 
         exec.emplace_back(read_executor, std::move(token_range));
     }
@@ -4099,7 +4101,7 @@ storage_proxy::query_partition_key_range_concurrent(storage_proxy::clock_type::t
 
     while (i != ranges.end()) {
         dht::partition_range& range = *i;
-        inet_address_vector_replica_set live_endpoints = get_live_sorted_endpoints(ks, end_token(range));
+        inet_address_vector_replica_set live_endpoints = get_live_sorted_endpoints_for_read(ks, end_token(range), cl, timeout);
         inet_address_vector_replica_set merged_preferred_replicas = preferred_replicas_for_range(*i);
         inet_address_vector_replica_set filtered_endpoints = filter_for_query(cl, ks, live_endpoints, merged_preferred_replicas, pcf);
         std::vector<dht::token_range> merged_ranges{to_token_range(range)};
@@ -4112,7 +4114,7 @@ storage_proxy::query_partition_key_range_concurrent(storage_proxy::clock_type::t
         {
             const auto current_range_preferred_replicas = preferred_replicas_for_range(*i);
             dht::partition_range& next_range = *i;
-            inet_address_vector_replica_set next_endpoints = get_live_sorted_endpoints(ks, end_token(next_range));
+            inet_address_vector_replica_set next_endpoints = get_live_sorted_endpoints_for_read(ks, end_token(next_range), cl, timeout);
             inet_address_vector_replica_set next_filtered_endpoints = filter_for_query(cl, ks, next_endpoints, current_range_preferred_replicas, pcf);
 
             // Origin has this to say here:
@@ -4639,10 +4641,65 @@ future<bool> storage_proxy::cas(schema_ptr schema, shared_ptr<cas_request> reque
     co_return condition_met;
 }
 
+bool storage_proxy::is_likely_alive(const gms::inet_address& ep, clock_type::time_point timeout) {
+    clock_type::time_point now = clock_type::now();
+    if (fbu::is_me(ep)) {
+        return true;
+    }
+    if (timeout <= now) {
+        return false;
+    }
+    netw::last_seen_info last_seen_ep = _messaging.get_last_seen_info_for(ep);
+    clock_type::duration time_left = timeout - now;
+    if (last_seen_ep.last_attempt == netw::last_seen_info::unknown || last_seen_ep.last_sent_without_response == netw::last_seen_info::unknown) {
+        return true;
+    }
+    clock_type::duration time_since_last_attempt = now - last_seen_ep.last_attempt;
+    if (time_since_last_attempt > time_left) {
+        return true;
+    }
+    clock_type::duration time_without_response = now - last_seen_ep.last_sent_without_response;
+    // Response time is estimated as the maximum of two values:
+    // - response time of the previous request
+    // - time without getting any response since the last time the coordinator started trying to communicate
+    //   with the replica
+    if (time_without_response > time_left) {
+        static constexpr int granularity = 10'000;
+        static thread_local std::uniform_int_distribution dice(0, granularity);
+        // There's always a small chance of probing the replica in case it came back to life.
+        int fail_chance = std::min<long>((time_without_response.count() - time_left.count()) * granularity / time_left.count(), granularity - 1);
+        int roll = dice(_urandom);
+        if (roll < fail_chance) {
+            slogger.debug("Overload protection: speculatively disqualifying {} as a candidate for a read request with {}ms left until timeout, "
+                    "since it hasn't responded for at least {}ms", ep, time_left.count(), time_without_response.count());
+            return false;
+        }
+    }
+    return true;
+}
+
 inet_address_vector_replica_set storage_proxy::get_live_endpoints(keyspace& ks, const dht::token& token) const {
     auto& rs = ks.get_replication_strategy();
     inet_address_vector_replica_set eps = rs.get_natural_endpoints_without_node_being_replaced(token);
     auto itend = boost::range::remove_if(eps, std::not1(std::bind1st(std::mem_fn(&gms::gossiper::is_alive), &gms::get_local_gossiper())));
+    eps.erase(itend, eps.end());
+    return eps;
+}
+
+inet_address_vector_replica_set storage_proxy::get_live_endpoints_for_read(keyspace& ks, const dht::token& token, db::consistency_level cl, clock_type::time_point timeout) {
+    auto& rs = ks.get_replication_strategy();
+    inet_address_vector_replica_set eps = rs.get_natural_endpoints_without_node_being_replaced(token);
+    auto itend = boost::range::remove_if(eps, [this, timeout, &gossiper = gms::get_local_gossiper()] (const gms::inet_address& ep) {
+        return !gossiper.is_alive(ep) || !is_likely_alive(ep, timeout);
+    });
+    // In the unlikely case when too many endpoints were disualified and it's impossible to fulfill
+    // the consistency level, bring back the endpoints which were "not likely to respond in time"
+    // and only remove the dead ones
+    if (!db::is_sufficient_live_nodes(cl, ks, eps.begin(), itend)) [[unlikely]] {
+        itend = std::remove_if(itend, eps.end(), [&gossiper = gms::get_local_gossiper()] (const gms::inet_address& ep) {
+            return !gossiper.is_alive(ep);
+        });
+    }
     eps.erase(itend, eps.end());
     return eps;
 }
@@ -4656,8 +4713,8 @@ void storage_proxy::sort_endpoints_by_proximity(inet_address_vector_replica_set&
     }
 }
 
-inet_address_vector_replica_set storage_proxy::get_live_sorted_endpoints(keyspace& ks, const dht::token& token) const {
-    auto eps = get_live_endpoints(ks, token);
+inet_address_vector_replica_set storage_proxy::get_live_sorted_endpoints_for_read(keyspace& ks, const dht::token& token, db::consistency_level cl, clock_type::time_point timeout) {
+    auto eps = get_live_endpoints_for_read(ks, token, cl, timeout);
     sort_endpoints_by_proximity(eps);
     return eps;
 }