This series presents a mechanism of disqualifying a node from read operations if it looks down. It's based on estimating how long ago the target node was still responsive - an as that time gets longer than request's timeout, the chance for dropping the request increases.

Spin-off from #8098, which also claimed to be able to estimate reliably if a replica is overloaded, which is considerably harder to do properly.