@psarna 2021-04-08
I'll also copy-paste grafana results from #8098, since they show the effects of the new failure detection:

> 1. A 3-node cluster is preloaded with some data, rf=3, cl=all
> 2. A quorum read workload starts - with custom USING TIMEOUT 1s CQL predicate; cl=quorum
> 3. One node is killed during the workload

> The red line in the middle indicates the point in which the cluster was upgraded to a version with the patches applied.

> What can be observed is that in the first case, right after one of the replicas is killed, we can observe a ~30s window of sharp increase in background reads along with accompanying read errors, as well as elevated number of digest reads, which is a result of speculative retry firing off.

> In the second case, background reads + read errors + digest reads are quickly translated to the new metrics - refused reads, which indicates that the node was disqualified as a candidate and the requests were rerouted to other nodes instead.

![with_and_without](https://user-images.githubusercontent.com/10433047/113270606-f3c0f600-92d9-11eb-8c45-d7c987a8f1c1.png)