@psarna 2021-04-08
Also, refs https://github.com/scylladb/scylla/issues/3746#issuecomment-762203969 - once #3746 is fixed, the accidental backpressure of background reads will not work anymore.