@psarna 2021-05-04
I pushed another iteration by the end of the day, but please do not review yet. I reworked the estimation calculations and added a fail-safe for the case when there are not enough replicas to satisfy the consistency level, but I haven't had time to run my test scenarios and update the docs yet.