@psarna 2021-05-05
v2:
 - reworked the calculations: the values gathered now are:
   - timestamp of a first sending attempt that didn't get its response
   - timestamp of the most recent sending attempt
   - round-trip time of the last successful request
 - added a fail-safe for the case when disqualifying a node also means that we won't be able to fulfill the consistency level - in that case, it's preferred to at least try contacting the replica, even if it's likely to be dead
 - amended the docs accordingly

Tests: manual, same scenarios as before, but for ONE, QUORUM and ALL consistency levels; testing with ALL allows checking if the fail-safe mechanism works