@psarna 2021-05-07
Example microbenchmark results below (`perf_simple_query --smp 1` ran multiple times).
master:
```
43087.77 tps ( 87.2 allocs/op,  20.2 tasks/op,   85885 insns/op)
43292.30 tps ( 87.2 allocs/op,  20.2 tasks/op,   85870 insns/op)
43154.16 tps ( 87.2 allocs/op,  20.2 tasks/op,   85871 insns/op)
43158.22 tps ( 87.2 allocs/op,  20.2 tasks/op,   85870 insns/op)
42427.62 tps ( 87.2 allocs/op,  20.2 tasks/op,   85868 insns/op)

median 43154.16 tps ( 87.2 allocs/op,  20.2 tasks/op,   85871 insns/op)
median absolute deviation: 66.39
maximum: 43292.30
minimum: 42427.62
```
this-series:
```
42220.20 tps ( 87.2 allocs/op,  20.2 tasks/op,   86184 insns/op)
42585.07 tps ( 87.2 allocs/op,  20.2 tasks/op,   86189 insns/op)
43336.23 tps ( 87.2 allocs/op,  20.2 tasks/op,   86227 insns/op)
42968.22 tps ( 87.2 allocs/op,  20.2 tasks/op,   86177 insns/op)
42978.42 tps ( 87.2 allocs/op,  20.2 tasks/op,   86196 insns/op)

median 42968.22 tps ( 87.2 allocs/op,  20.2 tasks/op,   86177 insns/op)
median absolute deviation: 368.01
maximum: 43336.23
minimum: 42220.20
```
Sounds like any change between master and this series is within the error margin.