@avikivity 2021-05-27
Usually the instructions counter is fairly accurate. It's unreasonable large, are you testing dev mode?

For stablest results, use perf_simple_query --smp 1 --task-quota-ms 10