@psarna 2021-05-27
I think it was release mode, but I have inline threshold set to 0 in order to avoid clang hanging, which may influence results. In any case, I'll just rerun with task-quota-ms 10 and see.