@psarna 2021-05-27
Bumped the inline threshold to 300 (600 is already enough to hang the compilation of row_level.cc, so I gave up higher numbers), here are the results:
```
Running test with config: {partitions=10000, concurrency=100, mode=read, frontend=cql, query_single_key=no, counters=no}
Creating 10000 partitions...
76410.47 tps ( 87.1 allocs/op,  20.1 tasks/op,   62250 insns/op)
75873.35 tps ( 87.1 allocs/op,  20.1 tasks/op,   62176 insns/op)
78467.92 tps ( 87.1 allocs/op,  20.1 tasks/op,   62209 insns/op)
80713.69 tps ( 87.1 allocs/op,  20.1 tasks/op,   62164 insns/op)
80561.72 tps ( 87.1 allocs/op,  20.1 tasks/op,   62174 insns/op)

median 78467.92 tps ( 87.1 allocs/op,  20.1 tasks/op,   62209 insns/op)
median absolute deviation: 2093.80
maximum: 80713.69
minimum: 75873.35
```
Is the instruction count still huge? I don't really have a reference