@avikivity 2021-05-27
The values I remember are 55k insn/query (that's high too).

But if it doesn't compile, how can we merge it?