@psarna 2021-05-27
The master branch doesn't compile for me since a couple of months - it's not the matter of this patch at all.