@denesb 2021-05-28
On Thu, 2021-05-27 at 07:38 -0700, Avi Kivity wrote:
> Usually the instructions counter is fairly accurate. It's
> unreasonable large, are you testing dev mode?
> For stablest results, use perf_simple_query --smp 1 --task-quota-ms
> 10
Should we make this the default, to avoid everybody running this with
the wrong arguments out-of-the-box?

> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub, or unsubscribe.