@psarna 2021-06-02
edit: I'll get acquainted with the new failure detector first, since it's merged, and then go back to this change - maybe it's no longer needed