@psarna 2021-06-02
edit2: no clash - the new direct failure detector still uses rather high granularity with default timeout of 20s, while this series targets requests with custom timeouts. As an example, ones with timeout set to 200ms will qualify nodes are dead considerably faster (~400ms), which makes the coordinator much less prone to getting overloaded with hanging background reads.