@avikivity 2021-06-24
Can the mechanism disqualify two nodes?

In general, we should disqualify at most (live nodes - requires_for_cl) nodes, otherwise we risk a false positive causing us to fail CL.