@avikivity 2021-06-27
> Also: I'm on vacation next week and there are merge conflicts already (and at a rate in which optimizations are applied to this area, new conflicts are guaranteed to accumulate), so I'll need to do a proper rebase of this series once I'm back.

The trick is to find problems with the optimizations so that they are delayed.