@psarna 2021-07-15
v3:
 - moved the detector one level down, the timestamps are now gathered in the messaging service, so all two-way messages contribute to updating the values
 - removed last remnants of gathering round-trip times, which were there in initial iterations, back when the series was both about failure detection and overload detection - now the detector only cares about detecting dead nodes, so round-trip times are irrelevant
 - updated docs accordingly
 
Tests: unit(release), manual(killing 1 node from a 3-node cluster during quorum reads)